requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
openai==1.12.0
python-multipart==0.0.9
pydantic==2.6.1
requests==2.31.0
aiohttp==3.9.3
SpeechRecognition==3.10.1
pydub==0.25.1
zeep==4.2.1
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0 